requirements.txt
streamlit
anthropic
requests
fpdf2
